chores: Reorganize bottom player
